split functions in ms_lexer.c
